+++ src/routes/examples/advanced/long-item/+page.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-    import Select from '$lib/Select.svelte';
+    import { Select } from '$lib';
 
     let collection = [
         {