+++ src/routes/examples/advanced/style-props/+page.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-    import Select from '$lib/Select.svelte';
+    import { Select } from '$lib';
 
     let items = ['One', 'Two', 'Three'];
     let listOpen = $state(false);