+++ src/routes/examples/events/clear/+page.svelte
@@ -1,6 +1,5 @@
 <script lang="ts">
-    import Select from '$lib/Select.svelte';
-    import type { SelectClearValue } from '$lib/types';
+    import { Select, type SelectClearValue } from '$lib';
 
     let items = [
         { value: 'one', label: 'One' },