+++ src/routes/examples/props/item-id/+page.svelte
@@ -1,13 +1,11 @@
 <script lang="ts">
-    import Select from '$lib/Select.svelte';
+    import { Select, type JustValue } from '$lib';
 
     let items = [
         { foo: 'value of One with value key foo', label: 'One' },
         { foo: 'value of Two with value key foo', label: 'Two' },
         { foo: 'value of Three with value key foo', label: 'Three' },
     ];
-
-    import type { JustValue } from '$lib/types';
 
     let justValue = $state<JustValue>(null);
 </script>